refactor(core): :recycle: rename profile file to get-profile-information
