+++ packages/core/src/api/public-api.ts
@@ -19,5 +19,5 @@
 export {default as authorize} from './authorize';
 export {default as authenticate} from './authenticate';
 export {default as getBrandingPreference} from './get-branding-preference';
-export {default as getProfileInformation} from './profile';
+export {default as getProfileInformation} from './get-profile-information';
 export {default as getBrandingPreferenceText} from './get-branding-preference-text';
